+++ aim/sdk/select.py
@@ -2,9 +2,10 @@ from typing import Optional
 
 from aim.engine.repo import AimRepo
 from aim.ql.grammar.statement import Statement
+from aim.ql.grammar.expression import Expression
 
 
-def select(search_statement: str, repo_path: Optional[str] = None):
+def select_metrics(search_statement: str, repo_path: Optional[str] = None):
     if repo_path is not None:
         repo = AimRepo(repo_full_path=repo_path)
     else:
@@ -25,3 +26,21 @@ def select(search_statement: str, repo_path: Optional[str] = None):
 
     return repo.select_metrics(statement_select,
                                statement_expr, default_expression)
+
+
+def select_runs(expression: Optional[str] = None,
+                repo_path: Optional[str] = None):
+    if repo_path is not None:
+        repo = AimRepo(repo_full_path=repo_path)
+    else:
+        repo = AimRepo.get_working_repo(mode=AimRepo.READING_MODE)
+
+    if not repo:
+        return None
+
+    if expression and 'run.archived' not in expression:
+        default_expression = 'run.archived is not True'
+    else:
+        default_expression = None
+
+    return repo.select_runs(expression, default_expression)