+++ examples/search/select_runs.py
@@ -1,0 +1,9 @@
+from aim.sdk import select_runs
+
+
+if __name__ == '__main__':
+    runs = select_runs((
+        'experiment == keras '
+        'and hparams.num_epochs >= 0.001 '
+    ))
+    print(runs)