+++ examples/search/statement.py
@@ -1,9 +1,0 @@
-from aim.ql.grammar import Statement
-
-
-if __name__ == '__main__':
-    st = Statement()
-    res = st.parse("""
-        loss, tf:loss if True
-    """)
-    print(res)