+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- Add ability to query runs via SDK (gorarakelyan)
+
 ## 2.0.12 Sep 12 2020
 - Update Session to handle exceptions gracefully (gorarakelyan)